requirements.txt
pyserial
google-generativeai
python-dotenv